requirements.txt
playwright~=1.43.0
pytest~=8.1.1
allure-pytest==2.13.2
requests>=2.25.1
python-dotenv~=1.0.1